Renamed IdentifierNode File to IdentifierField
